share setup code between tests and benchmarks
