+++ shotover-proxy/Cargo.toml
@@ -80,7 +80,6 @@ rusoto_signature = "0.46.0"
 
 [dev-dependencies]
 criterion = { version = "0.3", features = ["async_tokio", "html_reports"] }
-#proptest = "0.10.0"
 lazy_static = "1.4.0"
 ctrlc = "3.1.6"
 redis = {git = "https://github.com/benbromhead/redis-rs", branch = "parra_support", features = ["tokio-rt-core", "cluster"]}
@@ -92,9 +91,8 @@ hwaddr = "0.1.2"
 httparse = "1.3.0"
 threadpool = "1.0"
 num_cpus = "1.0"
-anyhow = "1.0.38"
 serial_test = "0.5.1"
-subprocess = "0.2.7"
+test-helpers = { path = "../test-helpers" }
 
 [[bench]]
 name = "redis_benches"
